Demote parts table to file attribute
Realized that ever since files came in in just one request, part URLs no longer have any reason to be added to the DB one at a time. We can just add them all at once, which also means it can just be represented as a newline-separated string in the files table on the DB instead of being its own bespoke table.