New chunking method; this one actually works
Found out about TransformStream, and it's exactly what I needed for bending the upload stream to my will. SetSizeChunkStream takes in a stream comprised of chunks of any size, and outputs a stream comprised of chunks of a _given_ size. So I use this as a filter to "condition" the chunks of the incoming stream into Discord's maximum file size. The TransformStream will merge and split the chunks it receives as necessary to fit them into chunks of the requested size.