Double down on anyness
The v8 release of better-sqlite3 changed the return type of .get and .all statements from any to unknown. Makes sense, but I hate enforcing conformity to interfaces in TypeScript so much I'm just not even going to bother.
And nevermind the redundant statement types, too! Infer them from the parameters of the function they're made for!